requirements.txt
accelerate==0.24.1
aiohttp==3.8.6
aiosignal==1.3.1
arrow==1.3.0
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
boto3==1.28.78
botocore==1.31.78
bravado==11.0.3
bravado-core==6.1.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==3.0.0
cryptography==41.0.5
datasets==2.14.6
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
docker==6.1.3
fabric==3.2.2
filelock==3.12.4
fqdn==1.5.1
frozenlist==1.4.0
fire==0.5.0
fsspec==2023.9.2
future==0.18.3
gin-config==0.5.0
gitdb==4.0.11
gitignore-parser==0.1.9
GitPython==3.1.40
google-auth==2.23.4
google-cloud==0.34.0
GradCache @ git+https://github.com/luyug/GradCache@e4234c9
huggingface-hub==0.19.4
idna==3.4
invoke==2.2.0
isoduration==20.11.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonlines==4.0.0
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kubernetes==28.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.7
mrunner @ git+https://gitlab.com/awarelab/mrunner@a195d266351e0b992a1e8d434f3c7ed4d7606b0b
mteb @ git+https://github.com/embeddings-benchmark/mteb@02e84b2fa8d147a86b4896d8e57e83f36285f5c7
multidict==6.0.4
multiprocess==0.70.15
munch==4.0.0
neptune==1.8.2
networkx==3.1
nltk==3.8.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
packaging==23.2
pandas==2.0.3
paramiko==3.3.1
path==16.7.1
path.py==12.5.0
peft==0.6.2
Pillow==9.3.0
psutil==5.9.6
pyarrow==14.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
PyNaCl==1.5.0
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==13.6.0
rpds-py==0.12.0
rsa==4.9
s3transfer==0.7.0
safetensors==0.3.3
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
swagger-spec-validator==3.0.3
sympy==1.12
termcolor==2.3.0
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.36.1
triton==2.1.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.18
webcolors==1.13
websocket-client==1.6.4
wrapt==1.15.0
xxhash==3.4.1
yarl==1.9.2
